+++ CHANGELOG.md
@@ -7,15 +7,17 @@ Documentation of release versions of `nacc-attribute-deriver`
 * Adds ability to derive type returned by a curation rule with tool to create table of rule types.
 * Adds enforcement of expected types when grabbing variables from the raw data
 * Adds rest of NP variables
-* Adds rest of UDS A1, A2 derived variables
-
+* Adds UDS A1, A2, A3, A4, and A5 derived variables **except** NACCFAM (A3) and and MEDS V1 (A4)
+    * Accounts for some forms being optional
+    * For A4, adds MEDS scope
+* Adds NACCNE4S and fixes NACCAPOE to account for ADC and ADGC genotypes not being the same
 
 ## 1.4.0
 
 * Add check for affiliated participants
 * Adds more NP variables
 * Refactors NP functionality - added `NPMapper` (mapping functions) and `NPFormWideEvaluator` (for variables like NACCBRNN that require looking at the entire NP form)
-* Refactors to consoildate redundant cross-sectional derived variables so UDS doesn't have to redefine them, namely for NP and genetics data
+* Refactors to consolidate redundant cross-sectional derived variables so UDS doesn't have to redefine them, namely for NP and genetics data
 * Fix `wmh` to not be required (for `SCANMRIScope.MRI_SBM`)
 
 ## 1.3.0